+++ examples/advanced_example_experiment.py
@@ -6,7 +6,7 @@ are parsed by a specific method. This avoids having one large "main" function wh
 """
 
 import numpy as np
-from seml.experiment import Experiment
+from seml import Experiment
 
 
 ex = Experiment()
