+++ src/seml/commands/configure.py
