+++ src/seml/evaluation.py
@@ -11,6 +11,7 @@ __all__ = ['get_results']
 
 def parse_jsonpickle(db_entry):
     import json
+
     import jsonpickle
     import jsonpickle.ext.numpy as jsonpickle_numpy
 
@@ -59,7 +60,9 @@ def get_results(
 
     """
     import functools
+
     import pandas as pd
+
     from seml.console import track
 
     if fields is None:
