+++ src/seml/utils/errors.py
